pyproject.toml
[tool.poetry]
name = "cert-generator"
version = "0.1.0"
description = ""
authors = ["litmus-zhang <abdulsalamlukmon9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.37.0"
openpyxl = "^3.1.5"
pillow = "^10.4.0"
mailerlite = "^0.1.6"
poetry-plugin-dotenv = "^2.2.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
